pyproject.toml
[tool.poetry]
name = "sibi-cooker"
version = "0.1.0"
description = ""
authors = ["Tim Sharpe <tim@bombasticmonkey.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
ricecooker = { git = "https://github.com/rodjek/ricecooker.git", branch = "py312" }
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
